eye_privacy/model/requirements.txt
fastapi>=0.68.0,<0.69.0
uvicorn>=0.15.0,<0.16.0
requests==2.25.1
numpy==1.22.3
Pillow==9.1.0